add json parsing middleware
